added a history page accessed by adding /history

+++ app.py
@@ -1,5 +1,6 @@
 from flask import Flask, render_template, request, redirect, url_for
 from flask_sqlalchemy import SQLAlchemy
+from datetime import datetime
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///todo.db'
@@ -8,6 +9,11 @@ db = SQLAlchemy(app)
 class Todo(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     content = db.Column(db.String(200), nullable=False)
+
+class DeletedTodo(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    content = db.Column(db.String(200), nullable=False)
+    deleted_at = db.Column(db.DateTime, default=datetime.utcnow)
 
 @app.route('/')
 def index():
@@ -25,9 +31,23 @@ def add():
 @app.route('/delete/<int:id>')
 def delete(id):
     todo_to_delete = Todo.query.get(id)
-    db.session.delete(todo_to_delete)
-    db.session.commit()
+
+    if todo_to_delete:
+        # Create a deleted todo record
+        deleted_todo = DeletedTodo(content=todo_to_delete.content)
+        db.session.add(deleted_todo)
+
+        # Delete the todo from the main table
+        db.session.delete(todo_to_delete)
+
+        db.session.commit()
+
     return redirect(url_for('index'))
+
+@app.route('/history')
+def history():
+    deleted_todos = DeletedTodo.query.all()
+    return render_template('history.html', deleted_todos=deleted_todos)
 
 if __name__ == '__main__':
     with app.app_context():
